handled food not found when delete food by id

+++ src/main/java/com/example/delivery/services/FoodService.java
@@ -21,6 +21,10 @@ public class FoodService {
     }
 
     public void deleteFoodById(Long id) {
+        if (!foodRepository.existsById(id)) {
+            throw new FoodNotFoundException(id);
+        }
+
         foodRepository.deleteById(id);
     }
 }
